+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v8.0.2-3] - 2026-09-24
+### Added
+- [375] Added some logging for error cases related to template expansion in CAS
+
 ## [v8.0.2-2] - 2026-09-21
 ### Fixed
 - [#373] Sign OIDC id tokens without an embedded `jwk` JOSE header